Update BuildFFUVM.ps1
- Implement basic Windows Update Catalog download caching

+++ FFUDevelopment/BuildFFUVM.ps1
@@ -312,6 +312,7 @@ param(
     [bool]$UpdateLatestMSRT,
     [bool]$UpdateEdge,
     [bool]$UpdateOneDrive,
+    [bool]$AllowUpdateCaching,
     [bool]$CopyPPKG,
     [bool]$CopyUnattend,
     [bool]$CopyAutopilot,
@@ -375,6 +376,7 @@ if (-not $VHDXPath) { $VHDXPath = "$VMPath\$VMName.vhdx" }
 if (-not $FFUCaptureLocation) { $FFUCaptureLocation = "$FFUDevelopmentPath\FFU" }
 if (-not $LogFile) { $LogFile = "$FFUDevelopmentPath\FFUDevelopment.log" }
 if (-not $KBPath) { $KBPath = "$FFUDevelopmentPath\KB" }
+if (-not $KBCachePath) { $KBCachePath = "$FFUDevelopmentPath\KBCache" }
 if (-not $DefenderPath) { $DefenderPath = "$AppsPath\Defender" }
 if (-not $MSRTPath) { $MSRTPath = "$AppsPath\MSRT" }
 if (-not $OneDrivePath) { $OneDrivePath = "$AppsPath\OneDrive" }
@@ -2260,6 +2262,46 @@ function Get-LatestWindowsKB {
     return $kbArticle
 }
 
+function Find-CachedKB {
+    param(
+        [string]$Link 
+    )
+
+    $fileHash = ($Link | Select-String -Pattern "(?<=^[^_]*_).*(?=\..*$)").Matches[0].Value
+    $file = [System.IO.FileInfo]::new((Join-Path -Path $KBCachePath -ChildPath ($Link -split '/')[-1]))
+
+    if ($file.Exists) {
+        if ((Get-FileHash -Path $file.FullName -Algorithm SHA1).Hash -eq $fileHash) {
+            Robocopy.exe $($KBCachePath) $($KBPath) $file.Name /E /R:5 /W:5 /J /Copy:DAT
+            return $file.Name
+        }
+    }
+
+    return ""
+}
+
+function Invoke-KBDownload {
+    param (
+        [string]$Link,
+        [string]$WindowsArch,
+        [string]$Path
+    )
+
+    if ($AllowUpdateCaching) {
+        $fileName = Find-CachedKB
+        if (![string]::IsNullOrEmpty($fileName)) {
+            Writelog "Found $Link for $WindowsArch in $KBCachePath"
+            Writelog "Returning $fileName"
+            return $fileName
+        }
+    }
+    Writelog "Downloading $Link for $WindowsArch to $Path"
+    Start-BitsTransferWithRetry -Source $link -Destination $Path
+    $fileName = ($link -split '/')[-1]
+    Writelog "Returning $fileName"
+    return $fileName
+}
+
 function Save-KB {
     [CmdletBinding()]
     param(
@@ -2286,19 +2328,13 @@ function Save-KB {
                     WriteLog "Link: $link matches security"
                     if ($WindowsArch -eq 'x64') {
                         if ($link -match 'securityhealthsetup_e1') {
-                            Writelog "Downloading $Link for $WindowsArch to $Path"
-                            Start-BitsTransferWithRetry -Source $link -Destination $Path
-                            $fileName = ($link -split '/')[-1]
-                            Writelog "Returning $fileName"
+                            $fileName = Invoke-KBDownload -Link $link -WindowsArch $WindowsArch -Path $Path
                             break
                         }
                     }
                     if ($WindowsArch -eq 'arm64') {
                         if ($link -match 'securityhealthsetup_25') {
-                            Writelog "Downloading $Link for $WindowsArch to $Path"
-                            Start-BitsTransferWithRetry -Source $link -Destination $Path
-                            $fileName = ($link -split '/')[-1]
-                            Writelog "Returning $fileName"
+                            $fileName = Invoke-KBDownload -Link $link -WindowsArch $WindowsArch -Path $Path
                             break
                         }
                     }
@@ -2308,10 +2344,7 @@ function Save-KB {
             if ($link -match 'x64' -or $link -match 'amd64') {
                 if($WindowsArch -is [array]) {
                     if ($link -match $WindowsArch[0] -or $link -match $WindowsArch[1]) {
-                        Writelog "Downloading $Link for $WindowsArch to $Path"
-                        Start-BitsTransferWithRetry -Source $link -Destination $Path
-                        $fileName = ($link -split '/')[-1]
-                        Writelog "Returning $fileName"
+                        $fileName = Invoke-KBDownload -Link $link -WindowsArch $WindowsArch -Path $Path
                         break
                     }
                 }
@@ -2319,10 +2352,7 @@ function Save-KB {
             }
             if ($link -match 'arm64') {
                 if ($WindowsArch -eq 'arm64') {
-                    Writelog "Downloading $Link for $WindowsArch to $Path"
-                    Start-BitsTransferWithRetry -Source $link -Destination $Path
-                    $fileName = ($link -split '/')[-1]
-                    Writelog "Returning $fileName"
+        	    $fileName = Invoke-KBDownload -Link $link -WindowsArch $WindowsArch -Path $Path
                     break
                 }
             }                
@@ -3523,6 +3553,11 @@ function Get-FFUEnvironment {
     Clear-InstallAppsandSysprep
     #Clean up $KBPath
     If (Test-Path -Path $KBPath) {
+         if ($AllowUpdateCaching) {
+            WriteLog "Copying $KBPath content to $KBCachePath for cache retrieval"
+            Robocopy.exe $($KBPath) $($KBCachePath) /E /Z /Copy:DAT
+            WriteLog "Copy complete"
+        }
         WriteLog "Removing $KBPath"
         Remove-Item -Path $KBPath -Recurse -Force -ErrorAction SilentlyContinue
         WriteLog 'Removal complete'
@@ -4125,9 +4160,14 @@ try {
             WriteLog 'This can take 10+ minutes depending on how old the media is and the size of the KB. Please be patient'
             Add-WindowsPackage -Path $WindowsPartition -PackagePath $KBPath -PreventPending | Out-Null
             WriteLog "KBs added to $WindowsPartition"
+	    if ($AllowUpdateCaching) {
+                WriteLog "Copying downloaded KBs from $KBPath to $KBCachePath"
+                Robocopy.exe $($KBPath) $($KBCachePath) /E /COPY:DAT /R:5 /W:5 /J
+                WriteLog "Copy complete"
+            }
             WriteLog "Removing $KBPath"
             Remove-Item -Path $KBPath -Recurse -Force | Out-Null
-	        WriteLog "Clean Up the WinSxS Folder"
+	    WriteLog "Clean Up the WinSxS Folder"
             Dism /Image:$WindowsPartition /Cleanup-Image /StartComponentCleanup /ResetBase | Out-Null
             WriteLog "Clean Up the WinSxS Folder completed"
         }
